58. Destroy Game Object
Added Bounce to the Left, Right, and Top Walls. Changed Bounce material to have a friction of 0. Created a Block script that destroys block when an object collides into it. Changed Block pixels per unit to 128, down from 256 (doubling the size). Changed the color of the sprite renderer for Block to Red.