59. Using Prefabs in Unity.
Created a Block prefab. Created a Blocks object in the hierarchy. Changed the snap settings for X and Y to 1. Set up a "Level 1".  Prefabbed the objects in our Level 1 that we would like to reuse in other scenes. Replaced the SceneLoader object in other scenes with the prefabbed version, and hooked up buttons to the prefabbed SceneLoader.