65. Count Breakable Blocks
Created Level GameObject prefab and script. Created a variable in Block.cs and Shape.cs to keep track of breakable objects. Created a methods in Level.cs to track the count of our breakable objects. We tell each breakable object (that has no parent shape) to add itself to the total at start and to remove itself from the total when destroyed.  Added Level prefab to scenes Level 1 and Level 2.