77. Smarter maxHits Logic
Changed logic so that max hits for a block is set to the number of sprites in the hitSprites array.
Gave Unbreakable Block a hitSprite to act as the default sprite. There are safeguards in Block.cs to prevent blocks that aren't breakable from taking damage.
Changed logic in Shape.cs so that the max hits for a shape is equal to the number of hitSprites for the first element of the blocks array.