78. Preventing Boring Ball Loops
Added a randomized velocity tweak to our Ball's Rigidbody2D.velocity on collision.
The intent is to prevent the ball from getting stuck in a loop.
While the course's solution works most of the time, I was still able to get stuck in a vertical loop on the very edge of the screen.