81. Autoplay for Playtesting
Added functionality in Paddle.cs and GameSession.cs to allow for an AutoPlay option, where the paddle will automatically map its x position to match the balls, allowing for automated playing.
Did a major refactor on Shape.cs. It no longer is tracking hits, and instead is damaging its children when it has a collision, and only breaks once all of its children have been destroyed.