Created MovementManager.
Movement Manager is a script that wihen placed on a GameObject, can specify interval-based movement.
As of now, there is functionality for movement on X axis, Y axis, and rotation around the Z axis.
There's also a method named MoveXRotateZRandomAction that, on every specified interval, will randomly decide whether to move left, right, or not at all, OR to rotate on Z axis to the left, right, or not at all.